remove CloseError
The final else case in LevelUP#close() corresponds to the status
'new', which is the state the database has after the constructor
has been called. Since we always call LevelUP#open() before
returning the database object, it's impossible for the user to
call close() while the database is in the state 'new'. Therefore
the CloseError() can never happen.