Align nextTick behavior with abstract-leveldown
By using `queueMicrotask()` in browsers, for a consistent experience
regardless of bundler (and of runtime in the future, when we drop
node 10). Webpack no longer shims node core modules.

In theory bundlers can now skip shimming `process`, were it not for
our `readable-stream` dependency which uses `process.nextTick()` at
the time of writing and has more reason to keep doing so. In modules
where the risk of timing issues (due to not having a common microtask
scheduler between modules) is low, like here in `levelup`, I prefer
the idiomatic `queueMicroTask()` function.

Lastly, maybe streams are not here to stay. Async iterators are
coming to the level ecosystem. Or maybe we'll switch to `streamx`
which is planning to use `queueMicrotask()` as well.